requirements.txt
pillow
numpy
torch
paddleocr
paddlepaddle
transformers
gunicorn